38. Makefile done. Norm done.

+++ srcs/prsng_vldtn/vldthn_map.c
@@ -6,13 +6,13 @@
 /*   By: bhatches <bhatches@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/16 21:48:32 by bhatches          #+#    #+#             */
-/*   Updated: 2021/04/19 12:26:44 by valery           ###   ########.fr       */
+/*   Updated: 2021/04/19 16:29:16 by bhatches         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "prsng_vldtn.h"
 
-int	vldthn_map_coordinates_space(t_cube3D *cube, int i, int j)
+int	vldthn_map_coordinates_space(t_cube3D *cube, size_t i, size_t j)
 {
 	if (i != 0 && j != 0 && j != ft_strlen(cube->map[i])
 		&& cube->map[i + 1] != NULL)
@@ -25,7 +25,7 @@ int	vldthn_map_coordinates_space(t_cube3D *cube, int i, int j)
 	return (0);
 }
 
-int	vldthn_map_coordinates_02NSWE(t_cube3D *cube, int i, int j)
+int	vldthn_map_coordinates_02NSWE(t_cube3D *cube, size_t i, size_t j)
 {
 	if (i != 0 && j != 0 && j != (ft_strlen(cube->map[i]) - 1)
 		&& cube->map[i + 1] != NULL)
@@ -46,7 +46,7 @@ int	vldthn_map_coordinates_02NSWE(t_cube3D *cube, int i, int j)
 	return (0);
 }
 
-int	vldthn_map_coordinates(t_cube3D *cube, int i, int j)
+int	vldthn_map_coordinates(t_cube3D *cube, size_t i, size_t j)
 {
 	if (ft_memchr(" 120NSWE", cube->map[i][j], 9) == NULL)
 		my_exit(11);
@@ -63,8 +63,8 @@ int	vldthn_map_coordinates(t_cube3D *cube, int i, int j)
 
 int	vldthn_map(t_cube3D *cube)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	while (cube->map[i] != NULL)
